[FEAT] Created the model Appointment and established his relations to the models User,Role and Store. Adjusted the datatypes on some models

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,13 +22,13 @@ return new class extends Migration
             $table -> string('city');
             $table -> string('postalCode');
             $table -> string('address');
-            $table -> integer('role_ID');
+            $table -> unsignedBigInteger('role_ID');
             $table -> boolean('isActive'); 
             $table->rememberToken();
 
 
 
-            $table -> foreign('role_ID') -> references('id') -> on('roles');
+            $table -> foreign('role_ID')->references('id')->on('roles');
             $table->timestamps();
         });
     }
